Ref #8 Remove duplicate code for logging. Created logging_utils.py
